backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "edge-tts>=7.2.7",
    "fastapi>=0.127.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.3",
]
